@@ -13,15 +13,12 @@ http://www.microsoft.com/resources/sharedsource/licensingbasics/publiclicense.ms
 ***************************************************************************/
 
 using System;
-using System.Drawing.Imaging;
 using System.IO;
 using System.Linq;
-using System.Runtime.InteropServices;
 using System.Text;
 using System.Xml.Linq;
 using DocumentFormat.OpenXml.Packaging;
 using OpenXmlPowerTools;
-using System.Collections.Generic;
 
 class WmlToHtmlConverterHelper
 {
@@ -84,59 +81,14 @@ class WmlToHtmlConverterHelper
                     {
                         ++imageCounter;
                         string extension = imageInfo.ContentType.Split('/')[1].ToLower();
-                        ImageFormat imageFormat = null;
-                        if (extension == "png")
-                            imageFormat = ImageFormat.Png;
-                        else if (extension == "gif")
-                            imageFormat = ImageFormat.Gif;
-                        else if (extension == "bmp")
-                            imageFormat = ImageFormat.Bmp;
-                        else if (extension == "jpeg")
-                            imageFormat = ImageFormat.Jpeg;
-                        else if (extension == "tiff")
-                        {
-                            // Convert tiff to gif.
-                            extension = "gif";
-                            imageFormat = ImageFormat.Gif;
-                        }
-                        else if (extension == "x-wmf")
-                        {
-                            extension = "wmf";
-                            imageFormat = ImageFormat.Wmf;
-                        }
-
-                        // If the image format isn't one that we expect, ignore it,
-                        // and don't return markup for the link.
-                        if (imageFormat == null)
-                            return null;
-
-                        string base64 = null;
-                        try
-                        {
-                            using (var imageStream = imageInfo.Bitmap)
-                            using (var ms = new MemoryStream())
-                            {
-                                imageStream.Seek(0, SeekOrigin.Begin);
-                                imageStream.CopyTo(ms);
-                                byte[] bytes = ms.ToArray();
-                                base64 = System.Convert.ToBase64String(bytes);
-                            }
-                        }
-                        catch (System.Runtime.InteropServices.ExternalException)
-                        {
-                            return null;
-                        }
-
-                        ImageCodecInfo codec = ImageCodecInfo.GetImageDecoders().First(c => c.FormatID == imageFormat.Guid);
-                        string mimeType = codec.MimeType;
-
-                        string imageSource = string.Format("data:{0};base64,{1}", mimeType, base64);
+                        String mimeType = $"image/{extension}";
+                        string base64 = Convert.ToBase64String(imageInfo.Bitmap);
+                        string imageSource = $"data:{mimeType};base64,{base64}";
 
                         XElement img = new XElement(Xhtml.img,
                             new XAttribute(NoNamespace.src, imageSource),
                             imageInfo.ImgStyleAttribute,
-                            imageInfo.AltText != null ?
-                                new XAttribute(NoNamespace.alt, imageInfo.AltText) : null);
+                            imageInfo.AltText != null ? new XAttribute(NoNamespace.alt, imageInfo.AltText) : null);
                         return img;
                     }
                 };
