@@ -374,43 +374,12 @@ AAsACwDBAgAAbCwAAAAA";
                                 localDirInfo.Create();
                             ++imageCounter;
                             string extension = imageInfo.ContentType.Split('/')[1].ToLower();
-                            ImageFormat imageFormat = null;
-                            if (extension == "png")
-                            {
-                                // Convert png to jpeg.
-                                extension = "gif";
-                                imageFormat = ImageFormat.Gif;
-                            }
-                            else if (extension == "gif")
-                                imageFormat = ImageFormat.Gif;
-                            else if (extension == "bmp")
-                                imageFormat = ImageFormat.Bmp;
-                            else if (extension == "jpeg")
-                                imageFormat = ImageFormat.Jpeg;
-                            else if (extension == "tiff")
-                            {
-                                // Convert tiff to gif.
-                                extension = "gif";
-                                imageFormat = ImageFormat.Gif;
-                            }
-                            else if (extension == "x-wmf")
-                            {
-                                extension = "wmf";
-                                imageFormat = ImageFormat.Wmf;
-                            }
-
-                            // If the image format isn't one that we expect, ignore it,
-                            // and don't return markup for the link.
-                            if (imageFormat == null)
-                                return null;
-
-                            string imageFileName = imageDirectoryName + "/image" + imageCounter.ToString() + "." + extension;
+                            string imageFileName = imageDirectoryName + "/image" + imageCounter + "." + extension;
                             try
                             {
                                 using (var fileStream = File.Create(imageFileName))
-                                using (var imageStream = imageInfo.Bitmap)
+                                using (var imageStream = new MemoryStream(imageInfo.Bitmap))
                                 {
-                                    imageStream.Seek(0, SeekOrigin.Begin);
                                     imageStream.CopyTo(fileStream);
                                 }
                             }
